+++ src/main/java/com/gokimpark/instaclone/domain/like/LikesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LikesRepository extends JpaRepository<Likes, Long> {
+public interface LikesRepository extends JpaRepository<Likes, Integer> {
 
     List<Likes> findAllByPost(Post post);
