+++ src/main/java/com/gokimpark/instaclone/web/post/PostShowDto.java
@@ -6,14 +6,14 @@ import lombok.Data;
 @Data
 public class PostShowDto {
 
-    private Long postId;
+    private Integer postId;
     private String username;
     private String imageUrl;
 
     private String likesCount;
 
     @Builder
-    public PostShowDto (Long postId, String username, String imageUrl, String likesCount){
+    public PostShowDto (Integer postId, String username, String imageUrl, String likesCount){
         this.postId = postId;
         this.username = username;
         this.imageUrl = imageUrl;
